Update version. Add Bank reg & log code
Added bank reg code as well as working code for proper logging. Yet to be perfectly modified but easy to update now. Banks are able to register & also know if there already exists a server for that Bank, which is also online. If not online, it simply updates the Port number and the server starts.